Create the first recipe
Create the index.html and README.md. Create the images and recipes directories.
Create the muffin.html and download the chocolatemuffin.jpg into the recipes and images respectively
Create unordered list of ingredients and ordered list of instructions along with heading and a description of the recipe on the muffin.html file.